+++ wow_dbc/src/wrath_tables/spell.rs
@@ -251,11 +251,11 @@ impl DbcTable for Spell {
             // effect_radius_index: int32[3]
             let effect_radius_index = crate::util::read_array_i32::<3>(chunk)?;
 
-            // effect_aura: EffectAura[3]
+            // effect_aura: AuraMod[3]
             let effect_aura = {
-                let mut arr = [EffectAura::default(); 3];
+                let mut arr = [AuraMod::default(); 3];
                 for i in arr.iter_mut() {
-                    *i = EffectAura::try_from(crate::util::read_i32_le(chunk)?)?;
+                    *i = AuraMod::try_from(crate::util::read_i32_le(chunk)?)?;
                 }
 
                 arr
@@ -737,7 +737,7 @@ impl DbcTable for Spell {
             }
 
 
-            // effect_aura: EffectAura[3]
+            // effect_aura: AuraMod[3]
             for i in row.effect_aura {
                 b.write_all(&(i.as_int() as i32).to_le_bytes())?;
             }
@@ -1016,7 +1016,7 @@ impl From<u16> for SpellKey {
 }
 
 #[derive(Debug, Clone, Copy, Ord, PartialOrd, Eq, PartialEq, Hash)]
-pub enum EffectAura {
+pub enum AuraMod {
     None,
     BindSight,
     ModPossess,
@@ -1336,7 +1336,7 @@ pub enum EffectAura {
     PeriodicHaste,
 }
 
-impl EffectAura {
+impl AuraMod {
     const fn from_value(value: i32) -> Option<Self> {
         Some(match value {
             0 => Self::None,
@@ -1661,15 +1661,15 @@ impl EffectAura {
     }
 }
 
-impl TryFrom<i32> for EffectAura {
+impl TryFrom<i32> for AuraMod {
     type Error = crate::InvalidEnumError;
     fn try_from(value: i32) -> Result<Self, Self::Error> {
-        Self::from_value(value).ok_or(crate::InvalidEnumError::new("EffectAura", value as i64))
+        Self::from_value(value).ok_or(crate::InvalidEnumError::new("AuraMod", value as i64))
     }
 
 }
 
-impl EffectAura {
+impl AuraMod {
     pub const fn as_int(&self) -> i32 {
         match self {
             Self::None => 0,
@@ -1995,7 +1995,7 @@ impl EffectAura {
 
 }
 
-impl Default for EffectAura {
+impl Default for AuraMod {
     fn default() -> Self {
         Self::None
     }
@@ -2066,7 +2066,7 @@ pub struct SpellRow {
     pub implicit_target_a: [i32; 3],
     pub implicit_target_b: [i32; 3],
     pub effect_radius_index: [i32; 3],
-    pub effect_aura: [EffectAura; 3],
+    pub effect_aura: [AuraMod; 3],
     pub effect_aura_period: [i32; 3],
     pub effect_amplitude: [f32; 3],
     pub effect_chain_targets: [i32; 3],